split warp bridge in half
